requirements.txt
beautifulsoup4==4.11.2
compoundfiles==0.3
compressed_rtf==1.0.6
dkimpy==1.1.4
dmarc==1.0.2
dnspython==2.4.0
fake_useragent==1.1.3
imgkit==1.2.3
ipython==8.5.0
maxminddb_geolite2==2018.703
msg_parser==1.2.0
#numpy==1.24.2
#pandas==1.5.3
#pycti==5.9.6
python_whois==0.8.0
Requests==2.31.0
tabulate==0.9.0
termcolor==2.3.0
validators==0.20.0
yara_python==4.2.0
lxml